+++ src/lexiforge/editorial/__init__.py
@@ -1,0 +1,11 @@
+from .changeset import ChangeSet, FileChange, ReleaseEligibilityImpact
+from .errors import EditorialError
+from .service import EditorialService
+
+__all__ = [
+    "ChangeSet",
+    "EditorialError",
+    "EditorialService",
+    "FileChange",
+    "ReleaseEligibilityImpact",
+]